Pull req fixes and formatting
